feat (vcs): github client refacto
