--- engine/vcs/repogithub/init.go
@@ -1,12 +1,0 @@
-package github
-
-var (
-	apiURL string
-	uiURL  string
-)
-
-// Init initializes github package
-func Init(apiurl, uiurl string) {
-	apiURL = apiurl
-	uiURL = uiurl
-}
